viewable icons and legal info
