--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "compile-hero.disable-compile-files-on-did-save-code": false
-}
