--- css/footer.css
@@ -1,24 +1,0 @@
-.footer {
-  padding: 5em 0 5em 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-}
-.footer__HBP {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-.footer__socials a {
-  justify-content: space-between;
-  padding-bottom: 2em;
-  align-items: center;
-  padding: 0 0.4rem;
-  margin-left: auto;
-  margin-right: auto;
-}
-.footer svg {
-  width: 150px;
-  height: 150px;
-}
