+++ css/mentors.css
@@ -1,0 +1,96 @@
+.mentor-card {
+  cursor: pointer;
+}
+
+.mentor-modal.show {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+}
+
+.mentor-modal .modal-dialog {
+  margin: auto;
+  max-width: unset;
+}
+
+.mentor-modal .modal-content {
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  padding: 32px;
+  border-radius: 24px;
+  border: none;
+  position: relative;
+}
+
+.mentor-modal .mentor-img {
+  width: 350px;
+  height: auto;
+  margin-bottom: 16px;
+  border-radius: 16px;
+}
+
+.mentor-modal .mentor-name {
+  font-size: 18px;
+  font-weight: bold;
+  margin-bottom: 4px;
+  margin-left: 8px;
+}
+
+.mentor-modal .mentor-pos {
+  font-size: 14px;
+  line-height: 16px;
+  margin-left: 8px;
+  margin-bottom: 0;
+}
+
+.mentor-modal__heading {
+  font-size: 18px;
+  font-weight: bold;
+  text-transform: uppercase;
+  margin: 8px 0 8px;
+}
+
+.mentor-modal__info {
+  margin-left: 48px;
+  min-width: 200px;
+}
+
+.mentor-modal__info-section {
+  margin-bottom: 24px;
+}
+
+.mentor-modal__info ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.mentor-modal__close {
+  position: absolute;
+  top: 24px;
+  right: 24px;
+  cursor: pointer;
+  height: 24px;
+  width: 24px;
+}
+
+.mentor-modal__close:before, .mentor-modal__close:after {
+  position: absolute;
+  left: 15px;
+  content: " ";
+  height: 24px;
+  width: 2px;
+  background-color: black;
+}
+
+.mentor-modal__close:before {
+  -webkit-transform: rotate(45deg);
+          transform: rotate(45deg);
+}
+
+.mentor-modal__close:after {
+  -webkit-transform: rotate(-45deg);
+          transform: rotate(-45deg);
+}
+/*# sourceMappingURL=mentors.css.map */
