+++ package.json
@@ -12,15 +12,9 @@
     "react": "^17.0.1",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
-    "react-scripts": "4.0.1",
-    "web-vitals": "^0.2.4", 
     "react-icons": "^3.11.0",
-    "gatsby": "^2.24.86",
-    "gatsby-plugin-google-analytics": "^2.3.19",
-    "gatsby-plugin-mailchimp": "^5.2.2",
-    "gatsby-plugin-react-helmet": "^3.3.14",
-    "gatsby-plugin-resolve-src": "^2.1.0",
-    "gatsby-plugin-sitemap": "^2.4.17"
+    "react-scripts": "4.0.1",
+    "web-vitals": "^0.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
