+++ src/styles/footer.scss
@@ -1,30 +1,30 @@
 .footer {
-    padding: 5em 0 5em 0;
-    display: flex;
-    overflow: hidden;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-
+    padding: 2em 0;
     &__socials {
-        .mobile-icons {
-            display: none;
-        }
+        //padding: 2rem;
+        text-align:center;
         a {
-        justify-content: space-between;
-        align-items: center;
-        padding: 0 0.5rem;
-        margin-left: auto;
-        margin-right: auto;
+        padding: 0 0.7rem;
         color: $dark-blue;
-    }}
+    }
+        
+    &__row {
+    flex-direction: row;
 
+    }
 
-    svg {
-        width: 150px;
-        height: 150px;
+    &__column {
+        text-align: center;
     }
 }
+    svg {
+        width: 220px;
+        height: 220px;
+    }
+
+}
+
+/*
 
 input {
     border: 0;
@@ -35,55 +35,52 @@ input {
 }
 
 .footer__form {
-    width: 100%;
+    //width: 100%;
     padding: 1em 0;
-
-    &-heading {
-        //text-transform: uppercase;
-        font-weight: bold;
+//text-align: left;
+    &__heading {
+        font-size: 18px;
+        font-weight: 500;
         color: $dark-blue;
-        text-align: left;
         margin-bottom: 1rem;
     }
 
-    &-msg {
+    &__msg {
         margin-top: 0.5rem;
     }
 
-    &-input {
+    &__input {
         border-radius: 10px;
         padding: 0 0.5rem;
     }
 
-    &-btn {
-        margin-top: 10px;
+    .primary-cta {
+        border:none;
+        border-radius: 5px;
+        padding: .4rem 0;;
     }
 }
-
-.footer__legal a {
+*/
+.footer__legal {
+    display: block;
+    text-align: center;
+    a {
     color: $orange;
 
+    }
     a:hover {
         font-weight: bold;
         text-decoration: none;
     }
 }
 
-.footer__HBP {
-    margin: 2em 0;
-    margin-bottom: 2em;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: space-evenly;
-}
 
-@media (min-width: 768px) {
-    .footer__HBP {
-        flex-direction: row;
+
+
+@media (max-width: 812px) {
+    .footer {
+        flex-direction: column;
+        justify-content: center;
     }
 
-    .logo {
-        margin-right: 6em;
-    }
 }
